Cleaned up support for durations in time units

+++ support/cas-server-support-jpa-ticket-registry/src/main/java/org/apereo/cas/ticket/registry/support/JpaLockingStrategy.java
@@ -42,7 +42,7 @@ public class JpaLockingStrategy implements LockingStrategy {
     private String uniqueId;
 
     /** Amount of time in seconds lock may be held. */
-    private int lockTimeout;
+    private long lockTimeout;
 
     /**
      * @param  id  Application identifier that identifies a row in the lock
@@ -71,7 +71,7 @@ public class JpaLockingStrategy implements LockingStrategy {
      *                  Use of a reasonable timeout facilitates recovery from node failures,
      *                  so setting to zero is discouraged.
      */
-    public void setLockTimeout(final int seconds) {
+    public void setLockTimeout(final long seconds) {
         if (seconds < 0) {
             throw new IllegalArgumentException("Lock timeout must be non-negative.");
         }
